hh-28,29: fix dto in domain and detekt
